third_party/blink/web_tests/compositing/text-input-element.html

<!doctype HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<input id="target" width=10 value="This is a long truncated text entry" style="font-size:40pt;"/>
<script>
if (window.internals) {
  internals.settings.setPreferCompositingToLCDTextEnabled(true);
}
function assertLayers(expected_count) {
  var json = JSON.parse(internals.layerTreeAsText(document));
  assert_equals(json.layers.length, expected_count, json);
}
test(() => {
  assertLayers(1);
}, "scrollable");
test(() => {
  target.value = "short";
  assertLayers(1);
}, "not scrollable");
</script>
