third_party/blink/web_tests/compositing/text-input-element-expected.txt

This is a testharness.js-based test.
FAIL scrollable assert_equals: [object Object] expected 3 but got 1
PASS not scrollable
Harness: the test ran to completion.
